requirements.txt
google-generativeai
flake8
